requirements.txt
beautifulsoup4==4.11.1
Flask==2.2.2
pymongo==4.2.0
python-dotenv==0.21.0
requests==2.22.0
waitress==2.1.2
